requirements.txt
setuptools==72.1.0
wheel==0.44.0
numpy==1.26.4
scipy==1.14.0
pandas==2.2.2
matplotlib==3.9.1.post1
scikit-learn==1.5.1
scikit-image==0.24.0
ffmpeg-python==0.2.0
opencv-python==4.10.0.84
torch==2.4.0
torchvision==0.19.0
torchmetrics==1.4.1
lightning==2.4.0
tqdm==4.66.5
transformers==4.44.0
diffusers==0.30.0
accelerate==0.33.0
xformers==0.0.27.post2
packaging==24.1
ninja==1.11.1.1
flash-attn==2.6.3
control==0.10.0
pytest==8.3.2
pillow==10.4.0
tensorboard==2.17.0
ftfy==6.2.3
jupyter-ai-magics==2.20.0
langchain==0.2.13
langchain-community==0.2.12
llama-index==0.10.65